testing_requirements.txt
django-nose==1.3
nose==1.3.4
coverage==3.7.1
mock==1.0.1
freezegun==0.3.3
